share parent sidebar
